Clean up redundant bool() casts and unnecessary dtype= in numpy calls
Agent-Logs-Url: https://github.com/motanova84/P-NP/sessions/9d96c659-a523-4239-a75d-155d76c72b2c

+++ pnp/echo_qcal/qcal_sync.py
@@ -50,7 +50,7 @@ class TemporalAlignmentVerifier:
         # Introducir una ligera no-uniformidad dependiente de la fase
         observed_freq[bin_index] += 10
         observed_freq[(bin_index + 2) % num_bins] -= 10
-        expected_freq = np.full(num_bins, total_count / num_bins, dtype=float)  # Distribución uniforme esperada
+        expected_freq = np.full(num_bins, total_count / num_bins)
         
         chi2_stat, p_value = stats.chisquare(observed_freq, expected_freq)
